+++ test/puppetlabs/pcp/client_test.clj
@@ -11,7 +11,6 @@
    (let [realized-future (future true)]
      (deref realized-future)                                ; make sure the future is realized
      (map->Client {:server "wss://localhost:8142/pcp/v1"
-                   :identity "pcp://the_identity/the_type"
                    :websocket-client ""
                    :websocket-connection (atom realized-future)
                    :handlers {}
@@ -49,19 +48,6 @@
              (dispatch-message client (message/make-message :message_type "foo"))))
       (is (= "default"
              (dispatch-message client (message/make-message :message_type "bar")))))))
-
-(def make-identity #'puppetlabs.pcp.client/make-identity)
-(deftest make-identity-test
-  (testing "It returns the correct identity"
-    (is (= "pcp://client01.example.com/test"
-           (make-identity "test-resources/ssl/certs/client01.example.com.pem" "test"))))
-  (testing "It returns the correct identity from a certificate chain"
-    (is (= "pcp://client01.example.com/test"
-           (make-identity "test-resources/ssl/certs/client01-chain.example.com.pem" "test"))))
-  (testing "It throws an exception when the certificate file is empty"
-    (is (thrown-with-msg? IllegalArgumentException
-                          #"test-resources/ssl/certs/client01-empty\.example\.com.pem must contain at least 1 certificate"
-                          (make-identity "test-resources/ssl/certs/client01-empty.example.com.pem" "test")))))
 
 (def make-connection #'puppetlabs.pcp.client/make-connection)
 (deftest make-connection-test